+++ src/Linna/Authorization/EnhancedUser.php
@@ -9,7 +9,9 @@
  */
 declare(strict_types=1);
 
-namespace Linna\Auth;
+namespace Linna\Authorization;
+
+use Linna\Authentication\User;
 
 /**
  * Enhanched User.