+++ src/Linna/Authorization/PermissionMapperInterface.php
@@ -9,7 +9,7 @@
  */
 declare(strict_types=1);
 
-namespace Linna\Auth;
+namespace Linna\Authorization;
 
 use Linna\DataMapper\DomainObjectInterface;
 use Linna\DataMapper\MapperInterface;